Some changes. It should work at least for the first cycle.

+++ ProjectCourse/Models/WorkoutPlan.cs
@@ -20,7 +20,7 @@ namespace ProjectCourse.Models
         public int Repetition { get; set; }
         public int WorkoutPlanSet { get; set; }
         public int Rest { get; set; }
-        public int WorkoutPlanWeight { get; set; }
+        public float WorkoutPlanWeight { get; set; }
     
         public virtual Plan Plan { get; set; }
         public virtual Workout Workout { get; set; }
